# packages/api-client/spec/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/orders": {
      "get": {
        "operationId": "OrdersController_list",
        "parameters": [
          {
            "name": "to",
            "required": false,
            "in": "query",
            "description": "ISO 8601 createdAt upper bound",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "description": "ISO 8601 createdAt lower bound",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "NEW",
                "CONFIRMED",
                "PACKED",
                "SHIPPED",
                "DELIVERED",
                "CANCELLED"
              ],
              "type": "string"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrdersListResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Некорректные query params",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiValidationErrorBodyDto"
                }
              }
            }
          },
          "401": {
            "description": "Не передан или недействителен Bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiAuthErrorBodyDto"
                }
              }
            }
          },
          "403": {
            "description": "Пользователь не является владельцем",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiAuthErrorBodyDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Owner-only список заказов с фильтрами и пагинацией",
        "tags": [
          "orders"
        ]
      },
      "post": {
        "description": "Публичный контракт ADR 0005: поля снимка покупателя обязательны; customerId — опционально для связанных негостевых заказов.",
        "operationId": "OrdersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Некорректное тело запроса",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiValidationErrorBodyDto"
                }
              }
            }
          },
          "404": {
            "description": "Покупатель, товар или строка склада не найдены",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiNotFoundErrorBodyDto"
                }
              }
            }
          },
          "409": {
            "description": "Нет на складе или товар неактивен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiBusinessConflictBodyDto"
                }
              }
            }
          }
        },
        "summary": "Создать гостевой заказ и зарезервировать склад",
        "tags": [
          "orders"
        ]
      }
    },
    "/orders/{id}": {
      "get": {
        "operationId": "OrdersController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Идентификатор заказа",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Не передан или недействителен Bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiAuthErrorBodyDto"
                }
              }
            }
          },
          "403": {
            "description": "Пользователь не является владельцем",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiAuthErrorBodyDto"
                }
              }
            }
          },
          "404": {
            "description": "Заказ не найден",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiNotFoundErrorBodyDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Получить заказ: снимок гостя, позиции, статусы, история",
        "tags": [
          "orders"
        ]
      }
    },
    "/orders/{id}/cancel": {
      "patch": {
        "operationId": "OrdersController_cancel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Идентификатор заказа",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManualOrderLifecycleTransitionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Некорректное тело запроса",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiValidationErrorBodyDto"
                }
              }
            }
          },
          "401": {
            "description": "Не передан или недействителен Bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiAuthErrorBodyDto"
                }
              }
            }
          },
          "403": {
            "description": "Пользователь не является владельцем",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiAuthErrorBodyDto"
                }
              }
            }
          },
          "404": {
            "description": "Заказ или склад не найдены",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiNotFoundErrorBodyDto"
                }
              }
            }
          },
          "409": {
            "description": "Недопустимый переход или нарушение инварианта склада",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiBusinessConflictBodyDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Отменить ещё не отгруженный заказ и снять резерв",
        "tags": [
          "orders"
        ]
      }
    },
    "/orders/{id}/invoice-sent": {
      "patch": {
        "operationId": "OrdersController_markInvoiceSent",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Идентификатор заказа",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManualOrderLifecycleTransitionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Некорректное тело запроса",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiValidationErrorBodyDto"
                }
              }
            }
          },
          "401": {
            "description": "Не передан или недействителен Bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiAuthErrorBodyDto"
                }
              }
            }
          },
          "403": {
            "description": "Пользователь не является владельцем",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiAuthErrorBodyDto"
                }
              }
            }
          },
          "404": {
            "description": "Заказ не найден",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiNotFoundErrorBodyDto"
                }
              }
            }
          },
          "409": {
            "description": "Недопустимый переход в жизненном цикле",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiBusinessConflictBodyDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Отметить, что счёт вручную отправлен",
        "tags": [
          "orders"
        ]
      }
    },
    "/orders/{id}/payment-confirmed": {
      "patch": {
        "operationId": "OrdersController_confirmPayment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Идентификатор заказа",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManualOrderLifecycleTransitionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Некорректное тело запроса",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiValidationErrorBodyDto"
                }
              }
            }
          },
          "401": {
            "description": "Не передан или недействителен Bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiAuthErrorBodyDto"
                }
              }
            }
          },
          "403": {
            "description": "Пользователь не является владельцем",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiAuthErrorBodyDto"
                }
              }
            }
          },
          "404": {
            "description": "Заказ не найден",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiNotFoundErrorBodyDto"
                }
              }
            }
          },
          "409": {
            "description": "Недопустимый переход в жизненном цикле",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiBusinessConflictBodyDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Подтвердить ручную оплату",
        "tags": [
          "orders"
        ]
      }
    },
...
    "/orders/{id}/notifications/resend": {
      "post": {
        "operationId": "OrdersController_resendNotification",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Идентификатор заказа",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Не передан или недействителен Bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiAuthErrorBodyDto"
                }
              }
            }
          },
          "403": {
            "description": "Пользователь не является владельцем",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiAuthErrorBodyDto"
                }
              }
            }
          },
          "404": {
            "description": "Заказ не найден",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiNotFoundErrorBodyDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Повторно отправить актуальное email-уведомление по заказу",
        "tags": [
          "orders"
        ]
      }
    },
    "/orders/{id}/status": {
      "patch": {
        "operationId": "OrdersController_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Идентификатор заказа",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrderStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Некорректное тело запроса",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiValidationErrorBodyDto"
                }
              }
            }
          },
          "401": {
            "description": "Не передан или недействителен Bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiAuthErrorBodyDto"
                }
              }
            }
          },
          "403": {
            "description": "Пользователь не является владельцем",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiAuthErrorBodyDto"
                }
              }
            }
          },
          "404": {
            "description": "Заказ или склад не найдены",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiNotFoundErrorBodyDto"
                }
              }
            }
          },
          "409": {
            "description": "Недопустимый переход или нарушение инварианта склада",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiBusinessConflictBodyDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Устаревший эндпоинт смены единого статуса",
        "tags": [
          "orders"
        ]
      }
    },
...
  },
  "info": {
    "title": "API Enot Tea",
    "description": "HTTP API для MVP: проверка работоспособности, каталог товаров и жизненный цикл заказа.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Локальная разработка"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ApiValidationErrorBodyDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "code": {
            "type": "string",
            "example": "VALIDATION_ERROR"
          },
          "message": {
            "type": "string"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationErrorField"
            }
          }
        },
        "required": [
          "statusCode",
          "code",
          "message"
        ]
      },
      "ApiAuthErrorBodyDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 401
          },
          "code": {
            "type": "string",
            "example": "AUTH_REQUIRED",
            "description": "Один из: AUTH_REQUIRED, AUTH_INVALID_TOKEN, AUTH_INVALID_CREDENTIALS, OWNER_ONLY"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "statusCode",
          "code",
          "message"
        ]
      },
      "ApiBusinessConflictBodyDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 409
          },
          "code": {
            "type": "string",
            "example": "INVALID_ORDER_STATUS_TRANSITION",
            "description": "Один из: INSUFFICIENT_STOCK, PRODUCT_INACTIVE, INVALID_ORDER_STATUS_TRANSITION, INVENTORY_INVARIANT_VIOLATION, PRODUCT_DUPLICATE"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "required": [
          "statusCode",
          "code",
          "message"
        ]
      },
...
      "OrdersListResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderListItemDto"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/OrderPaginationDto"
          }
        },
        "required": [
          "items",
          "pagination"
        ]
      },
...
      "ApiNotFoundErrorBodyDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 404
          },
          "message": {
            "type": "string",
            "example": "Заказ orderId=order-1 не найден."
          },
          "error": {
            "type": "string",
            "example": "Не найдено"
          }
        },
        "required": [
          "statusCode",
          "message",
          "error"
        ]
      },
...
      "CreateOrderDto": {
        "type": "object",
        "properties": {
          "customerId": {
            "type": "string",
            "description": "Необязательный связанный id покупателя для негостевых заказов; для гостевого checkout не указывайте."
          },
          "customerFullName": {
            "type": "string",
            "example": "Иван Иванов",
            "description": "Снимок ФИО гостя/покупателя, хранимый в заказе."
          },
          "customerEmail": {
            "type": "string",
            "format": "email",
            "example": "customer@example.com",
            "description": "Снимок email гостя/покупателя, хранимый в заказе."
          },
          "customerPhone": {
            "type": "string",
            "example": "+995 555 010 010",
            "description": "Необязательный снимок телефона."
          },
          "shippingAddress": {
            "type": "string",
            "example": "Тбилиси, ул. Руставели, 1",
            "description": "Снимок адреса доставки, хранимый в заказе."
          },
          "items": {
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateOrderItemDto"
            }
          }
        },
        "required": [
          "customerFullName",
          "customerEmail",
          "shippingAddress",
          "items"
        ]
      },
      "ManualOrderLifecycleTransitionDto": {
        "type": "object",
        "properties": {
          "comment": {
            "type": "string",
            "maxLength": 500
          }
        }
      },
      "UpdateOrderStatusDto": {
        "type": "object",
        "properties": {
          "toStatus": {
            "type": "string",
            "enum": [
              "CONFIRMED",
              "PACKED",
              "SHIPPED",
              "DELIVERED"
            ],
            "description": "Целевой статус; CANCELLED недопустим (используйте /cancel)."
          },
          "comment": {
            "type": "string",
            "maxLength": 500
          }
        },
        "required": [
          "toStatus"
        ]
      },
...
    }
  }
}